services/ocr/tesseract/requirements.txt
flask==3.0.2
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
python-magic==0.4.27
gunicorn==21.2.0